Merge progress in development branch to main
A lot has happened an i know that this is a bad commit
message but it's all i have for now. I've add pypaystack
module to the project root. I've added more implementation
to the customer dashboard and used HTMX for dynamic content
on the dashboard. added transaction model to the finance app